Move listeners into `tcp.go`.
