+++ service/tcp.go
@@ -15,6 +15,7 @@
 package service
 
 import (
+	"bufio"
 	"bytes"
 	"container/list"
 	"context"
@@ -33,6 +34,7 @@ import (
 	onet "github.com/Jigsaw-Code/outline-ss-server/net"
 	"github.com/Jigsaw-Code/outline-ss-server/service/metrics"
 	logging "github.com/op/go-logging"
+	proxyproto "github.com/pires/go-proxyproto"
 	"github.com/shadowsocks/go-shadowsocks2/socks"
 )
 
@@ -229,6 +231,46 @@ func (c *clientStreamConn) ClientAddr() net.Addr {
 		return c.clientAddr
 	}
 	return c.StreamConn.RemoteAddr()
+}
+
+// StreamListener wraps a [net.Listener].
+type StreamListener struct {
+	net.Listener
+}
+
+// Accept waits for and returns the next incoming connection.
+func (l *StreamListener) Accept() (IClientStreamConn, error) {
+	c, err := l.Listener.Accept()
+	if err != nil {
+		return nil, err
+	}
+	c.(*net.TCPConn).SetKeepAlive(true)
+	return &clientStreamConn{StreamConn: c.(transport.StreamConn)}, nil
+}
+
+// ProxyListener wraps a [StreamListener] and fetches the source of the connection from the PROXY
+// protocol header string. See https://www.haproxy.org/download/1.8/doc/proxy-protocol.txt.
+type ProxyListener struct {
+	StreamListener
+}
+
+// Accept waits for the next incoming connection, parses the client IP from the PROXY protocol
+// header, and adds it to the connection.
+func (l *ProxyListener) Accept() (IClientStreamConn, error) {
+	conn, err := l.StreamListener.Accept()
+	if err != nil {
+		return nil, err
+	}
+	r := bufio.NewReader(conn)
+	h, err := proxyproto.Read(r)
+	if err == proxyproto.ErrNoProxyProtocol {
+		logger.Warningf("Received connection from %v without proxy header.", conn.RemoteAddr())
+		return conn, nil
+	}
+	if err != nil {
+		return nil, fmt.Errorf("error parsing proxy header: %v", err)
+	}
+	return &clientStreamConn{StreamConn: conn, clientAddr: h.SourceAddr}, nil
 }
 
 type StreamAccepter func() (IClientStreamConn, error)
